requirements.txt
altair
google-cloud-bigquery
google-cloud-storage
pandas
streamlit
